refactor definition placement in different files
